fix(issue#33): 효과음이 들리지 않는 문제 해결
GamePlayScene에 있는 SoundManager를 타이틀에 복사했음.
기존 SoundManager의 BGM 재생 부분은 MusicManager라는 새로운 게임 오브젝트와 스크립트를 만들어 해결함.
다만 새로운 문제가 생김: 게임 플레이 단계에서 몇번 반복 재생을 하다 갑자기 타이틀 스크린 배경음악을 재생하는 문제가 발생함.
원인을 모르겠음. 새로운 이슈 생성 예정.

close: issue#33

close: issue#33
